FLASK_RC1/requirements.txt
Flask==1.1.2
Jinja2==2.11.2
MarkupSafe==1.1.1
Werkzeug==1.0.1
beautifulsoup4==4.9.0
gunicorn==20.0.4
itsdangerous==1.1.0
requests==2.23.0
six==1.14.0
Flask-SQLAlchemy==2.3.2
Flask-CSV==1.2.0
psycopg2==2.8.5
python-dotenv==0.13.0
